pyproject.toml
[project]
name = "realtime-cdc-streaming-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker>=37.4.0",
    "jupyter==1.1.1",
    "psycopg2-binary==2.9.10",
    "python-decouple==3.8",
    "rich>=14.0.0",
    "sqlalchemy>=2.0.41",
]
